+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - 2024-??-?? 4.0.57
     - fix check_refs in BoxcarFeature
     - support Expert mode in plugin fields
+    - add Session Count to Measurement metadata
+    - consistently handle plugin metadata
 - 2024-03-23 4.0.56
     - updated ColumnFileParser with wasatch.CSVLoader updates
     - changing Preset auto-clears Dark and Reference